pyproject.toml
[tool.poetry]
name = "ocr-backend"
version = "0.1.0"
description = "OCR Backend API using FastAPI"
authors = ["Pongsakon <pongsakon.kaew@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
python-dotenv = "^1.1.0"
pydantic-settings = "^2.8.0"
python-multipart = "^0.0.20"
slowapi = "^0.1.9"
aiofiles = "^24.1.0"
coloredlogs = "^15.0.1"
pydantic = {extras = ["email"], version = "^2.11.0"}
pillow = "^11.1.0"
httpx = "^0.28.1"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-cov = "^6.1.0"
pytest-asyncio = "^0.26.0"
pytest-mock = "^3.14.0"
black = "^25.1.0"
flake8 = "^7.2.0"
mypy = "^1.15.0"
isort = "^6.0.0"
ipython = "^9.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=app --cov-report=term-missing"
asyncio_mode = "auto" 